requirements.txt
python-ffmpeg
fire
google-api-core
google-auth
google-cloud-core
google-cloud-speech
google-cloud-storage
google-cloud-texttospeech
google-cloud-translate
google-crc32c
google-resumable-media
googleapis-common-protos
pydub
moviepy
python-dotenv
dataclasses-json
jinja2
dotted-dict
azure-cognitiveservices-speech
